Revert "[#17] move from double to BigDecimal"
This reverts commit 994cbc2a1da39022e1313f6778e6f050cfe3fe8c.

+++ src/main/java/com/kubawach/nfs/core/model/Product.java
@@ -1,7 +1,5 @@
 package com.kubawach.nfs.core.model;
 
-import java.math.BigDecimal;
-
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
@@ -13,5 +11,5 @@ public class Product {
         return new Product(source.getConcentration());
     }
 
-    private BigDecimal concentration;
+    private double concentration;
 }
